# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Arch Linux Security Tracker API
  version: '1'
  description: Public CVE, package, AVG, published advisory and synchronization reads; scoped token writes.
    JSON request bodies are limited to 65536 bytes. Public records include notes. Public GET/HEAD responses
    support conditional reads; authenticated drafts and mutations are no-store. PATCH requires the exact strong
    ETag from a fresh detail GET in If-Match. Tokens have one or more scopes and expire after 90 days. See api-v1.md,
    api-catalog.md and api-workflow.md.
servers:
  - {url: /api/v1}
security: []
paths:
  /cves:
    get:
      operationId: listCVEs
      summary: List and filter CVEs, including orphan records
      description: Filters combine with AND. q searches literal substrings in identifier, description and notes.
        Lexical identifier pagination is not a snapshot or synchronization checkpoint.
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/CVEAfter'}
        - {$ref: '#/components/parameters/PackageFilter'}
        - name: orphan
          in: query
          schema:
            type: string
            enum: ['true', 'false']
        - name: severity
          in: query
          schema: {$ref: '#/components/schemas/Severity'}
        - name: q
          in: query
          schema: {type: string}
        - {$ref: '#/components/parameters/IfNoneMatch'}
      responses:
        '200': {$ref: '#/components/responses/CVEPage'}
        default: {$ref: '#/components/responses/Error'}
        '304': {$ref: '#/components/responses/NotModified'}
    post:
      operationId: createCVE
      summary: Create a public CVE without an AVG
      description: 'Requires cves:create scope. Immediately publishes all fields including notes. Duplicate
        identifier: 409 without merging.'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CVECreate'}
      responses:
        '201': {$ref: '#/components/responses/CVECreated'}
        default: {$ref: '#/components/responses/Error'}
  /cves/{name}:
    parameters:
      - {$ref: '#/components/parameters/CVEName'}
    get:
      operationId: getCVE
      summary: Read one public CVE
      parameters:
        - {$ref: '#/components/parameters/IfNoneMatch'}
      responses:
        '200': {$ref: '#/components/responses/CVE'}
        default: {$ref: '#/components/responses/Error'}
        '304': {$ref: '#/components/responses/NotModified'}
    patch:
      operationId: updateCVE
      summary: Update reviewed CVE fields
      description: Requires cves:update scope. Advisory-linked records additionally require Security Team membership.
        Refreshes linked group severity and scheduled advisory types.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CVEPatch'}
      responses:
        '200': {$ref: '#/components/responses/CVE'}
        default: {$ref: '#/components/responses/Error'}
        '412': {$ref: '#/components/responses/Error'}
        '428': {$ref: '#/components/responses/Error'}
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Personal token created through the CSRF-protected /tokens browser page. Select one or more scopes:
        cves:create, cves:update, groups:create, groups:update, intake:create, or advisories:write. Expires after 90 days;
        owner activity and role are checked on every request. Reporter or higher is required; advisories:write
        requires Security Team or administrator. Browser cookies do not authorize API operations.'
  responses:
    Error:
      description: 'JSON error. 400 bad_request/invalid_json/invalid_idempotency_key; 401 unauthorized (Bearer challenge); 403 forbidden;
        404 not_found; 405 method_not_allowed; 409 already_exists/already_grouped/advisory_exists/already_published/group_not_fixed/concurrent_creation/idempotency_conflict;
        412 precondition_failed; 413 payload_too_large; 415 unsupported_media_type; 422 validation_error; 428
        precondition_required; 500 internal_error. All errors use Cache-Control: no-store. fields is present
        when individual field errors are available.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotModified:
      description: Unchanged public representation. No response body.
      headers:
        ETag: {$ref: '#/components/headers/ETag'}
    CVE:
      description: CVE representation.
      headers:
        ETag: {$ref: '#/components/headers/ETag'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CVE'}
    CVEPage:
      description: CVEPage representation.
      headers:
        ETag: {$ref: '#/components/headers/ETag'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CVEPage'}
...
    CVECreated:
      description: CVE created.
      headers:
        Location: {$ref: '#/components/headers/Location'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CVE'}
...
  schemas:
    CVEName:
      type: string
      pattern: ^CVE-[0-9]{4}-[0-9]{4,}$
      maxLength: 64
      description: Syntax validation only; does not establish CVE assignment.
    CVE:
      type: object
      required: [name, type, severity, vector, description, notes, references, groups, packages, created, updated,
        cvss]
      properties:
        name: {type: string}
        type: {type: string}
        severity: {$ref: '#/components/schemas/Severity'}
        vector: {$ref: '#/components/schemas/Vector'}
        description: {type: string}
        notes: {type: string, description: Public notes.}
        references:
          type: array
          items: {type: string}
        groups:
          type: array
          uniqueItems: true
          items: {type: string}
          description: Sorted AVG identifiers; empty for orphan CVEs.
        packages:
          type: array
          uniqueItems: true
          items: {type: string}
          description: Sorted package names linked through AVGs.
        created: {type: string, format: date-time}
        updated:
          type: string
          format: date-time
          description: Stored UTC modification timestamp. Use /changes, not timestamps, for synchronization.
        cvss: {$ref: '#/components/schemas/CVSS'}
    Severity:
      type: string
      default: unknown
      enum: [unknown, critical, high, medium, low]
    Vector:
      type: string
      default: unknown
      enum: [unknown, remote, local]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            fields:
              type: object
              additionalProperties:
                type: array
                items: {type: string}
    IssueType:
      type: string
      enum:
        - unknown
        - access restriction bypass
        - arbitrary code execution
        - arbitrary command execution
        - arbitrary file overwrite
        - arbitrary filesystem access
        - arbitrary file upload
        - authentication bypass
        - certificate verification bypass
        - content spoofing
        - cross-site request forgery
        - cross-site scripting
        - denial of service
        - directory traversal
        - incorrect calculation
        - information disclosure
        - insufficient validation
        - man-in-the-middle
        - open redirect
        - private key recovery
        - privilege escalation
        - proxy injection
        - same-origin policy bypass
        - sandbox escape
        - session hijacking
        - signature forgery
        - silent downgrade
        - sql injection
        - time alteration
        - url request injection
        - xml external entity injection
    References:
      type: array
      default: []
      description: HTTP(S) URLs without whitespace or control characters. Deduplicated in first-occurrence
        order. Newline-joined total after deduplication must not exceed 4096 characters. URLs are not fetched.
      items: {type: string, format: uri, pattern: '^[Hh][Tt][Tt][Pp][Ss]?://', maxLength: 4096}
    Text: {type: string, maxLength: 4096, description: Valid Unicode scalar values.}
    CVSS:
      type: object
      nullable: true
      additionalProperties: false
      required: [version, score, vector, source]
      description: Source assessment, independent of Arch severity. Null clears it. Vector syntax and prefix
        are checked; score is not recalculated.
      properties:
        version:
          type: string
          enum: ['3.0', '3.1', '4.0']
        score: {type: number, minimum: 0, maximum: 10, multipleOf: 0.1}
        vector: {type: string, maxLength: 256, pattern: '^CVSS:(3\.0|3\.1|4\.0)(/[A-Z][A-Z0-9]*:[A-Za-z0-9]+)+$'}
        source: {type: string, format: uri, maxLength: 2048, description: HTTP(S) source URL without whitespace/control
            characters.}
    CVECreate:
      type: object
      additionalProperties: false
      required: [name]
      description: Omitted type/severity/vector default to unknown, text to empty, references to [], CVSS to
        null. Null is accepted only for cvss.
      properties:
        name: {$ref: '#/components/schemas/CVEName'}
        type: {$ref: '#/components/schemas/IssueType'}
        severity: {$ref: '#/components/schemas/Severity'}
        vector: {$ref: '#/components/schemas/Vector'}
        description: {$ref: '#/components/schemas/Text'}
        notes: {$ref: '#/components/schemas/Text'}
        references: {$ref: '#/components/schemas/References'}
        cvss: {$ref: '#/components/schemas/CVSS'}
    CVEPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      description: Only supplied fields change. Identifier cannot be renamed. Null is accepted only for cvss.
      properties:
        type: {$ref: '#/components/schemas/IssueType'}
        severity: {$ref: '#/components/schemas/Severity'}
        vector: {$ref: '#/components/schemas/Vector'}
        description: {$ref: '#/components/schemas/Text'}
        notes: {$ref: '#/components/schemas/Text'}
        references: {$ref: '#/components/schemas/References'}
        cvss: {$ref: '#/components/schemas/CVSS'}
...
    CVEPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/CVE'}
        next_cursor: {type: string, nullable: true}
...
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
...
    CVEAfter:
      name: after
      in: query
      schema: {$ref: '#/components/schemas/CVEName'}
      description: Exclusive lexical CVE identifier cursor.
...
    CVEName:
      name: name
      in: path
      required: true
      schema: {$ref: '#/components/schemas/CVEName'}
...
    PackageFilter:
      name: package
      in: query
      schema: {type: string}
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema: {type: string}
      description: ETag from an earlier response; a matching representation returns 304.
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string}
      description: 'Exact quoted strong ETag from a fresh GET. Missing: 428; stale or wildcard: 412.'
  headers:
    ETag:
      description: Strong SHA-256 validator of the complete JSON representation.
      schema: {type: string}
    Location:
      description: Relative detail URL of the new record.
      schema: {type: string}